backend/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.31.0
SpeechRecognition==3.10.1
pyaudio==0.2.14
Flask-CORS==4.0.1